+++ src/components/Sidebar/index.tsx
@@ -1,9 +1,9 @@
 import * as Dialog from '@radix-ui/react-dialog'
-import NewPageModal from './Modal/NewPageModal'
-import NavList from './NavList'
-import Menu from './Menu'
+import NewPageModal from '../Modal/NewPageModal'
+import NavList from '../NavList'
+import Menu from '../Menu'
 
-import IconPlusRounded from '../assets/icons/IconPlusRounded'
+import IconPlusRounded from '@/assets/icons/IconPlusRounded'
 
 const Sidebar = () => {
   return (
